Rename component files
Closes #3
